+++ tests/roundtrip_gltf.py
@@ -16,6 +16,8 @@ import bpy
 import os
 import sys
 
+bpy.ops.preferences.addon_enable(module="io_hubs_addon")
+
 try:
     argv = sys.argv
     if "--" in argv:
@@ -26,7 +28,26 @@ try:
     filepath = argv[0]
 
     bpy.ops.object.select_all(action='SELECT')
-    bpy.ops.object.delete(use_global=False)
+
+    for collection in bpy.data.collections:
+        for object in collection.objects:
+            collection.objects.unlink(object)
+
+    for bpy_data_iter in (
+        bpy.data.objects,
+        bpy.data.meshes,
+        bpy.data.lights,
+        bpy.data.cameras,
+        bpy.data.armatures,
+        bpy.data.actions,
+        bpy.data.images,
+        bpy.data.lightprobes,
+        bpy.data.materials,
+        bpy.data.shape_keys,
+        bpy.data.textures
+    ):
+        for id_data in bpy_data_iter:
+            bpy_data_iter.remove(id_data)
 
     bpy.ops.import_scene.gltf(filepath=argv[0])
 
@@ -41,12 +62,18 @@ try:
     output_dir = os.path.join(path_parts[0], argv[1])
     if not os.path.exists(output_dir):
         os.makedirs(output_dir)
+    if '--use-variants' in argv:
+        bpy.context.preferences.addons['io_scene_gltf2'].preferences.KHR_materials_variants_ui = True
     if '--no-sample-anim' in argv:
         bpy.ops.export_scene.gltf(export_format=export_format, filepath=os.path.join(
-            output_dir, path_parts[1]), export_force_sampling=False)
+            output_dir, path_parts[1]), export_force_sampling=False, export_cameras=True)
+    elif '--use-original-specular' in argv:
+        bpy.ops.export_scene.gltf(export_format=export_format, filepath=os.path.join(
+            output_dir, path_parts[1]), export_original_specular=True, export_cameras=True)
     else:
         bpy.ops.export_scene.gltf(
-            export_format=export_format, filepath=os.path.join(output_dir, path_parts[1]))
+            export_format=export_format, filepath=os.path.join(output_dir, path_parts[1]), export_cameras=True)
+
 except Exception as err:
     print(err, file=sys.stderr)
     sys.exit(1)
