+++ tests/test/tests/hemisphere-light.js
@@ -1,0 +1,27 @@
+const fs = require('fs');
+const path = require('path')
+const assert = require('assert');
+const utils = require('../utils.js');
+
+module.exports = {
+    description: 'can export hemisphere-light',
+    test: outDirPath => {
+        let gltfPath = path.resolve(outDirPath, 'hemisphere-light.gltf');
+        const asset = JSON.parse(fs.readFileSync(gltfPath));
+
+        assert.strictEqual(asset.extensionsUsed.includes('MOZ_hubs_components'), true);
+        assert.strictEqual(utils.checkExtensionAdded(asset, 'MOZ_hubs_components'), true);
+
+        const node = asset.nodes[0];
+        assert.strictEqual(utils.checkExtensionAdded(node, 'MOZ_hubs_components'), true);
+
+        const ext = node.extensions['MOZ_hubs_components'];
+        assert.deepStrictEqual(ext, {
+            "hemisphere-light": {
+                "skyColor": "#0cff00",
+                "groundColor": "#0cff00",
+                "intensity": 1
+            }
+        });
+    }
+};
