+++ tests/test/tests/personal-space-invader.js
@@ -1,0 +1,27 @@
+const fs = require('fs');
+const path = require('path')
+const assert = require('assert');
+const utils = require('../utils.js');
+
+module.exports = {
+    description: 'can export personal-space-invader',
+    test: outDirPath => {
+        let gltfPath = path.resolve(outDirPath, 'personal-space-invader.gltf');
+        const asset = JSON.parse(fs.readFileSync(gltfPath));
+
+        assert.strictEqual(asset.extensionsUsed.includes('MOZ_hubs_components'), true);
+        assert.strictEqual(utils.checkExtensionAdded(asset, 'MOZ_hubs_components'), true);
+
+        const node = asset.nodes[0];
+        assert.strictEqual(utils.checkExtensionAdded(node, 'MOZ_hubs_components'), true);
+
+        const ext = node.extensions['MOZ_hubs_components'];
+        assert.deepStrictEqual(ext, {
+            "personal-space-invader": {
+                "radius": 0.10000000149011612,
+                "useMaterial": false,
+                "invadingOpacity": 0.30000001192092896
+            }
+        });
+    }
+};
