+++ tests/test/tests/reflection-probe.js
@@ -1,0 +1,29 @@
+const fs = require('fs');
+const path = require('path')
+const assert = require('assert');
+const utils = require('../utils.js');
+
+module.exports = {
+    description: 'can export reflection-probe',
+    test: outDirPath => {
+        let gltfPath = path.resolve(outDirPath, 'reflection-probe.gltf');
+        const asset = JSON.parse(fs.readFileSync(gltfPath));
+
+        assert.strictEqual(asset.extensionsUsed.includes('MOZ_hubs_components'), true);
+        assert.strictEqual(utils.checkExtensionAdded(asset, 'MOZ_hubs_components'), true);
+
+        const node = asset.nodes[0];
+        assert.strictEqual(utils.checkExtensionAdded(node, 'MOZ_hubs_components'), true);
+
+        const ext = node.extensions['MOZ_hubs_components'];
+        assert.deepStrictEqual(ext, {
+            "reflection-probe": {
+                "size": 2.5,
+                "envMapTexture": {
+                    "__mhc_link_type": "texture",
+                    "index": 0
+                }
+            }
+        });
+    }
+};
