+++ tests/test/tests/scale-audio-feedback.js
@@ -1,0 +1,26 @@
+const fs = require('fs');
+const path = require('path')
+const assert = require('assert');
+const utils = require('../utils.js');
+
+module.exports = {
+    description: 'can export scale-audio-feedback',
+    test: outDirPath => {
+        let gltfPath = path.resolve(outDirPath, 'scale-audio-feedback.gltf');
+        const asset = JSON.parse(fs.readFileSync(gltfPath));
+
+        assert.strictEqual(asset.extensionsUsed.includes('MOZ_hubs_components'), true);
+        assert.strictEqual(utils.checkExtensionAdded(asset, 'MOZ_hubs_components'), true);
+
+        const node = asset.nodes[0];
+        assert.strictEqual(utils.checkExtensionAdded(node, 'MOZ_hubs_components'), true);
+
+        const ext = node.extensions['MOZ_hubs_components'];
+        assert.deepStrictEqual(ext, {
+            "scale-audio-feedback": {
+                "minScale": 1,
+                "maxScale": 1.5
+            }
+        });
+    }
+};
