+++ tests/test/utils.js
@@ -89,7 +89,7 @@ function materialWithName(gltf, name) {
   };
 }
 
-exports.utils = {
+module.exports = {
   UUID_REGEX,
   blenderFileToGltf,
   blenderRoundtripGltf,
